Allow connection messenger to ignore response (#70)
* Allow connection messenger to ignore response

* individualize item updates

* npm package version for shared models

* formatting

* remove log

+++ src/managers/connection-manager/connection-manager.ts
@@ -44,9 +44,13 @@ export class ConnectionManager implements IConnectionManager {
     }
 
     async deleteConnection(connectionId: string): Promise<void> {
-        this._logger.debug(`deleting connection ${connectionId}`);
-
         const expenseId = await this._connectionDao.getExpenseIdForConnection(connectionId);
+
+        this._logger.debug(`deleting connection ${connectionId} for expense ${expenseId}`);
+        if (!connectionId || !expenseId) {
+            return;
+        }
+
         await this._connectionDao.delete({ connectionId, expenseId });
     }
